# savicar-api/docs/swagger.yaml
basePath: /
definitions:
  badencar-api_internal_domain_services.Service:
    properties:
      code:
        type: string
      description:
        type: string
      hours_quantity:
        type: number
      id_order:
        type: integer
      id_service:
        type: integer
      total_value:
        type: number
      unit_value:
        type: number
    type: object
...
  operationalcosts.OperationalCost:
    properties:
      amount:
        type: number
      created_at:
        type: string
      description:
        type: string
      id_cost:
        type: integer
      id_cost_category:
        type: integer
      id_order:
        type: integer
      recurrence:
        type: integer
      reference_date:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Service Order Images API
  title: Badencar API
  version: "1.0"
paths:
...
  /operational-costs:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/operationalcosts.OperationalCost'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all operational costs
      tags:
      - operational-costs
    post:
      consumes:
      - application/json
      parameters:
      - description: Payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/operationalcosts.OperationalCost'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/operationalcosts.OperationalCost'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create an operational cost
      tags:
      - operational-costs
  /operational-costs/{id}:
    delete:
      parameters:
      - description: Operational Cost ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete an operational cost
      tags:
      - operational-costs
    get:
      parameters:
      - description: Operational Cost ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/operationalcosts.OperationalCost'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get an operational cost by ID
      tags:
      - operational-costs
    put:
      consumes:
      - application/json
      parameters:
      - description: Operational Cost ID
        in: path
        name: id
        required: true
        type: integer
      - description: Payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/operationalcosts.OperationalCost'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/operationalcosts.OperationalCost'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update an operational cost
      tags:
      - operational-costs
...
  /services:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/badencar-api_internal_domain_services.Service'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all services
      tags:
      - services
    post:
      consumes:
      - application/json
      parameters:
      - description: Payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/badencar-api_internal_domain_services.Service'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/badencar-api_internal_domain_services.Service'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a service
      tags:
      - services
  /services/{id}:
    delete:
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a service
      tags:
      - services
    get:
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/badencar-api_internal_domain_services.Service'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a service by ID
      tags:
      - services
    put:
      consumes:
      - application/json
      parameters:
      - description: Service ID
        in: path
        name: id
        required: true
        type: integer
      - description: Payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/badencar-api_internal_domain_services.Service'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/badencar-api_internal_domain_services.Service'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a service
      tags:
      - services
...
swagger: "2.0"