# savicar-api/internal/infrastructure/http/handler/serviceappointmentresource_handler.go
package handler

import (
	"net/http"
	"strconv"

	"github.com/gin-gonic/gin"

	sarsvc "savicar-api/internal/application/serviceappointmentresource"
	"savicar-api/internal/domain/serviceappointmentresource"
)

type ServiceAppointmentResourceHandler struct {
	svc *sarsvc.Service
}

func NewServiceAppointmentResourceHandler(svc *sarsvc.Service) *ServiceAppointmentResourceHandler {
	return &ServiceAppointmentResourceHandler{svc: svc}
}

func (h *ServiceAppointmentResourceHandler) RegisterRoutes(r gin.IRouter) {
	g := r.Group("/appointment-resources")
	g.GET("/by-appointment/:id", h.getByAppointment)
	g.GET("/:id", h.getByID)
	g.POST("", h.create)
	g.PUT("/:id", h.update)
	g.DELETE("/:id", h.delete)
}

func (h *ServiceAppointmentResourceHandler) getByAppointment(c *gin.Context) {
	id, err := strconv.Atoi(c.Param("id"))
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": "invalid id"})
		return
	}
	items, err := h.svc.GetByAppointmentID(c.Request.Context(), id)
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}
	if items == nil {
		items = []serviceappointmentresource.ServiceAppointmentResource{}
	}
	c.JSON(http.StatusOK, items)
}

func (h *ServiceAppointmentResourceHandler) getByID(c *gin.Context) {
	id, err := strconv.Atoi(c.Param("id"))
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": "invalid id"})
		return
	}
	item, err := h.svc.GetByID(c.Request.Context(), id)
	if err != nil {
		c.JSON(http.StatusNotFound, gin.H{"error": err.Error()})
		return
	}
	c.JSON(http.StatusOK, item)
}

func (h *ServiceAppointmentResourceHandler) create(c *gin.Context) {
	var item serviceappointmentresource.ServiceAppointmentResource
	if err := c.ShouldBindJSON(&item); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": "invalid body"})
		return
	}
	if err := h.svc.Create(c.Request.Context(), &item); err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}
	c.JSON(http.StatusCreated, item)
}

func (h *ServiceAppointmentResourceHandler) update(c *gin.Context) {
	id, err := strconv.Atoi(c.Param("id"))
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": "invalid id"})
		return
	}
	var item serviceappointmentresource.ServiceAppointmentResource
	if err := c.ShouldBindJSON(&item); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": "invalid body"})
		return
	}
	item.IDServiceAppointmentResource = id
	if err := h.svc.Update(c.Request.Context(), &item); err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}
	c.JSON(http.StatusOK, item)
}

func (h *ServiceAppointmentResourceHandler) delete(c *gin.Context) {
	id, err := strconv.Atoi(c.Param("id"))
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": "invalid id"})
		return
	}
	if err := h.svc.Delete(c.Request.Context(), id); err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}
	c.Status(http.StatusNoContent)
}